Don't directly modify webdriver prototype
Uses `addCommand` instead of directly modifying the webdriverIO
prototype. The client has additional steps besides adding the method as
a property, so this will keep more in sync with changes to how commands
are added (hopefully requiring little or no updates on this end for
compatibility).

Directly modifying the protoype also appears to cause memory/performance
issues when using `babel-register` in the webdriverIO browser test
runner.